% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeny.R
\name{add_cherry_to_tip}
\alias{add_cherry_to_tip}
\title{Adds taxon as sister to a tip.}
\usage{
add_cherry_to_tip(tree, tip, new_tips, position = 0.5)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{phylo}.}

\item{tip}{A single character variable containing the exact tip name of the
sister froup of the new taxon. Must be present as a tree$tip.label.}

\item{new_tips}{A single character variable or a string of characters
containing the tip names to add.}

\item{position}{A numeric value between\code{0} and \code{1} to set where on the edge
leading to the sister tip the new splitting node should be placed.}
}
\value{
A Phylogenetic tree of class \code{phylo}.
}
\description{
Adds taxon as sister to a tip.
}
\examples{
set.seed(1)
# library(tibble)
# library(phytools)
# library(ape)
library(dplyr)

tree <- phytools::pbtree(n=10,scale=1)

tree$tip.label <- LETTERS[ape::Ntip(tree):1]
# plot(tree)

par(mfrow=c(1,4))

tip_colors_orig <- tibble::tibble(tip = tree$tip.label, color = NA) \%>\%
  mutate(color = case_when(grepl("^N", tip) ~ "red",
                           tip == "C" ~ "green",
                           TRUE ~ "darkgreen"))

ape::plot.phylo(tree,
           mar=c(0.1,0.1,1.1,0.1),
           cex=1.5,
           tip.color = tip_colors_orig$color)
title(main = paste0("original"), cex.main=1.5)

for(i in c(.2,.5,.8)){ # c(.1,.25,.5,.75,.9)
  tree_new1 <- add_cherry_to_tip(tree = tree,
                                 tip = "C",
                                 new_tips = "N1",
                                 position = i)


  # get colors
  tip_colors <- tibble::tibble(tip = tree_new1$tip.label, color = NA) \%>\%
    mutate(color = case_when(grepl("^N", tip) ~ "red",
                             tip == "C" ~ "green",
                             TRUE ~ "darkgreen"))

  ape::plot.phylo(tree_new1,
             mar=c(0.1,0.1,1.1,0.1),
             cex=1.5,
             tip.color = tip_colors$color)

  title(main = paste0("pos.= ", i), cex.main=1.5)
}

ape::plot.phylo(tree,
           mar=c(0.1,0.1,1.1,0.1),
           cex=1.5,
           tip.color = tip_colors_orig$color)
title(main = paste0("original"), cex.main=1.5)

for(i in c(.2,.5,.8)){ # c(.1,.25,.5,.75,.9)
tree_new1 <- add_cherry_to_tip(tree = tree,
                                 tip = "C",
                                 new_tips = c("N1", "N2", "N3", "N4"),
                                 position = i)

  # get colors
  tip_colors <- tibble::tibble(tip = tree_new1$tip.label, color = NA) \%>\%
    mutate(color = case_when(grepl("^N", tip) ~ "red",
                             tip == "C" ~ "green",
                             TRUE ~ "darkgreen"))

  ape::plot.phylo(tree_new1,
             mar=c(0.1,0.1,1.1,0.1),
             cex=1.5,
             tip.color = tip_colors$color)

  title(main = paste0("pos.= ", i), cex.main=1.5)
}

par(mar=c(5.1,4.1,4.1,2.1))

}
