% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{add_leading_zeros}
\alias{add_leading_zeros}
\title{Add leading zero to a vector of numbers.}
\usage{
add_leading_zeros(numbers, length)
}
\arguments{
\item{numbers}{A vector of \code{numeric} or \code{character} values.}

\item{length}{A single \code{numeric} value defining the desired length of each
\code{character} of the resulting vector.}
}
\value{
A vector of \code{characters}.
}
\description{
Add leading zero to a vector of numbers.
}
\examples{
# add leading zeros to numbers
add_leading_zeros(numbers = c(1, 2, 3),
                   length = 4)

# add leading zeros to characters
add_leading_zeros(numbers = c("01", "2", "003"),
                   length = 4)
}
