% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{print_progress}
\alias{print_progress}
\title{Show loop progress in percent.}
\usage{
print_progress(current, end)
}
\arguments{
\item{current}{\code{Numeric} value of current loop iteration.}

\item{end}{\code{Numeric} value of loop end value.}
}
\description{
Show loop progress in percent.
}
\examples{
n = 100
for(i in 1:n){
  print_progress(i, n)
}
}
