% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_gbif}
\alias{taxonomy_gbif}
\title{Get taxize Gbif taxonomy data for a list of taxonomic genera.}
\usage{
taxonomy_gbif(genera, taxon = "Insecta")
}
\arguments{
\item{genera}{A vector of \code{character} values for taxonomic genera.}

\item{taxon}{A \code{character} string for taxonomic context to search. Can be
\code{Animalia}, \code{Arthropoda}, or \code{Insecta}.}
}
\value{
A data frame in the \code{tibble} format with the colums \code{genus}, \code{class},
\code{order}, and \code{family}.
}
\description{
Get taxize Gbif taxonomy data for a list of taxonomic genera.
}
\examples{
# add leading zeros to numbers

taxonomy_gbif(c("Caenorhabditis", "Drosophila", "Xenopus"), taxon = "Animalia")

taxonomy_gbif(c("Argiope", "Limulus", "Folsomia"), taxon = "Arthropoda")

taxonomy_gbif(c("Calopteryx", "Forficula", "Drosophila"), taxon = "Insecta")

}
